# contracts/estate.openapi.yaml
openapi: 3.0.3
info:
  title: Paymax × Spotlight — Estate Super-App Module API
  version: 1.0.0
  description: >
    Estate vertical (Blocks 29–46): dues/payments, meetings, tasks, maintenance,
    facilities, announcements, emergencies, documents, vendors, property
    management, AI meeting notes, finance dashboard, admin panel, notifications,
    reports, and member settings.

    Conventions:
      - All monetary amounts are integers in minor units (kobo). Never floats.
      - Money mutations require an Idempotency-Key header and post a balanced,
        tier-checked ledger entry (see /estate/dues/{id}/pay).
      - Auth: Bearer JWT issued by Supabase Auth. Estate scope is derived
        server-side from estate_residents; admin-only routes also require
        role = estate_admin.
    Module spec — mirrors contracts/visitor.openapi.yaml. Served under /api/v1.
servers:
  - url: /api/v1
    description: Estate module routes (Next.js route handlers)
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound: { description: Resource not found or not in caller's estate }
...
  schemas:
...
    Meeting:
      type: object
      properties:
        id: { type: string, format: uuid }
        estateId: { type: string, format: uuid }
        title: { type: string }
        agenda: { type: string, nullable: true }
        mode: { type: string, enum: [physical, virtual, hybrid] }
        location: { type: string, nullable: true }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [scheduled, live, ended, cancelled] }
        rsvpCounts: { type: object, additionalProperties: { type: integer } }
        myRsvp: { type: string, enum: [yes, no, maybe], nullable: true }

    # ── Tasks (Block 31) ──
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        estateId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        assigneeId: { type: string, format: uuid, nullable: true }
        assigneeName: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [todo, in_progress, done] }
        createdAt: { type: string, format: date-time }

    # ── Maintenance / repairs (Block 32) ──
...
security:
  - bearerAuth: []
paths:
...
  /estate/meetings/{id}/rsvp:
    post:
      operationId: rsvpMeeting
      summary: Set the caller's RSVP (upsert)
      tags: [Meetings]
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [response], properties: { response: { type: string, enum: [yes, no, maybe] } } }
      responses:
        '200': { description: RSVP recorded, content: { application/json: { schema: { $ref: '#/components/schemas/Meeting' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /estate/tasks/{id}:
    get:
      operationId: getTask
      summary: Task detail
      tags: [Tasks]
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200': { description: Task, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...